Support million-body broadphase stress scenes

+++ src/container.h
@@ -43,7 +43,7 @@
 #define b3Array_CreateN( a, n )                                                                                                  \
 	do                                                                                                                           \
 	{                                                                                                                            \
-		( a ).data = ( n ) > 0 ? b3GrowAlloc( NULL, 0, ( n ) * sizeof( *( a ).data ) ) : NULL;         \
+		( a ).data = ( n ) > 0 ? b3GrowAlloc( NULL, 0, (size_t)( n ) * sizeof( *( a ).data ) ) : NULL;                  \
 		( a ).count = 0;                                                                                                         \
 		( a ).capacity = ( n );                                                                                                  \
 	}                                                                                                                            \
@@ -64,9 +64,9 @@
 	{                                                                                                                            \
 		if ( ( a ).capacity < n )                                                                                                \
 		{                                                                                                                        \
-			int oldSize = ( a ).capacity * sizeof( *( a ).data );                                                                \
-			int newSize = ( n ) * sizeof( *( a ).data );                                                                         \
-			( a ).data = b3GrowAlloc( ( a ).data, oldSize, newSize );                                  \
+			size_t oldSize = (size_t)( a ).capacity * sizeof( *( a ).data );                                                     \
+			size_t newSize = (size_t)( n ) * sizeof( *( a ).data );                                                              \
+			( a ).data = b3GrowAlloc( ( a ).data, oldSize, newSize );                                                           \
 			( a ).capacity = ( n );                                                                                              \
 		}                                                                                                                        \
 	}                                                                                                                            \
@@ -86,10 +86,10 @@
 	{                                                                                                                            \
 		if ( ( a ).count >= ( a ).capacity )                                                                                     \
 		{                                                                                                                        \
-			int oldSize = ( a ).capacity * sizeof( *( a ).data );                                                                \
+			size_t oldSize = (size_t)( a ).capacity * sizeof( *( a ).data );                                                     \
 			int newCapacity = ( a ).capacity == 0 ? 8 : 2 * ( a ).capacity;                                                      \
-			int newSize = newCapacity * sizeof( *( a ).data );                                                                   \
-			( a ).data = b3GrowAlloc( ( a ).data, oldSize, newSize );                                  \
+			size_t newSize = (size_t)newCapacity * sizeof( *( a ).data );                                                        \
+			( a ).data = b3GrowAlloc( ( a ).data, oldSize, newSize );                                                           \
 			( a ).capacity = newCapacity;                                                                                        \
 		}                                                                                                                        \
 		( a ).data[( a ).count++] = ( value );                                                                                   \
@@ -119,12 +119,12 @@
 		{                                                                                                                        \
 			int req = ( a ).count + _n;                                                                                          \
 			int newCapacity = req > 2 ? req + ( req >> 1 ) : 8;                                                                  \
-			int oldSize = ( a ).capacity * sizeof( *( a ).data );                                                                \
-			int newSize = newCapacity * sizeof( *( a ).data );                                                                   \
-			( a ).data = b3GrowAlloc( ( a ).data, oldSize, newSize );                                  \
+			size_t oldSize = (size_t)( a ).capacity * sizeof( *( a ).data );                                                     \
+			size_t newSize = (size_t)newCapacity * sizeof( *( a ).data );                                                        \
+			( a ).data = b3GrowAlloc( ( a ).data, oldSize, newSize );                                                           \
 			( a ).capacity = newCapacity;                                                                                        \
 		}                                                                                                                        \
-		memcpy( ( a ).data + ( a ).count, ( src ), _n * sizeof( *( a ).data ) );                                                 \
+		memcpy( ( a ).data + ( a ).count, ( src ), (size_t)_n * sizeof( *( a ).data ) );                                         \
 		( a ).count += _n;                                                                                                       \
 	}                                                                                                                            \
 	while ( 0 )
@@ -144,14 +144,14 @@
 // B3_NULL_INDEX, otherwise it returns the index of the last element (which is now out of bounds).
 #define b3Array_RemoveSwap( a, index ) b3RemoveHelper( ( a ).data, &( a ).count, ( index ), sizeof( *( a ).data ) )
 
-B3_INLINE void* b3EmplaceHelper( void** data, int* count, int* capacity, int elem_size )
+B3_INLINE void* b3EmplaceHelper( void** data, int* count, int* capacity, size_t elem_size )
 {
 	if ( *count >= *capacity )
 	{
 		int oldCapacity = *capacity;
-		int oldSize = oldCapacity * elem_size;
+		size_t oldSize = (size_t)oldCapacity * elem_size;
 		int newCapacity = ( oldCapacity == 0 ? 16 : 2 * oldCapacity );
-		int newSize = newCapacity * elem_size;
+		size_t newSize = (size_t)newCapacity * elem_size;
 		*data = b3GrowAlloc( *data, oldSize, newSize );
 		*capacity = newCapacity;
 	}
